added usePasswordValidation hook
