+++ src/src_V2/hooks/index.js
@@ -1,1 +1,3 @@
+export * from './useTimer';
 export * from './useHoverTooltip';
+export * from './usePasswordValidation';
