+++ src/src_V2/modules/auth/components/verifyEmail/VerifyEmail.jsx
@@ -3,8 +3,8 @@ import { Divider, Link, Typography } from '@mui/material';
 
 import AuthInfoWrapper from '../authInfoWrapper/AuthInfoWrapper';
 
+import { useTimer } from '../../../../hooks';
 import styles from './VerifyEmail.module.css';
-import useTimer from '../../../../hooks/useTimer';
 import verifyIllustration from '../../../../assets/verifyIllustration.svg';
 
 // Temporary fix, no date formatting utils
