+++ src/src_V2/modules/auth/restorePassword/SetNewPassword.jsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { Formik, Form } from 'formik';
 
 import Button from '../../../components/formUI/Button';
@@ -7,38 +6,23 @@ import AuthPageWrapper from '../components/authWrapper/authPageWrapper';
 import PasswordValidator from '../components/passwordValidator/PasswordValidator';
 
 import styles from './RestorePassword.module.css';
-import { ERROR_MESSAGES, getPasswordStrengthLevel } from '../../../utils';
+import { usePasswordValidation } from '../../../hooks';
 import { setNewPasswordInitialValues } from '../../../utils/schemas/auth/auth';
 
 export default function SetNewPassword() {
-  const [passwordStrengthLevel, setPasswordStrengthLevel] = useState(0);
-
-  const validatePasswordHandler = (password) => {
-    const strengthLevel = getPasswordStrengthLevel(password);
-    setPasswordStrengthLevel(strengthLevel);
-    if (!password) return ERROR_MESSAGES.required;
-    if (strengthLevel < 31) return ERROR_MESSAGES.password;
-  };
-
-  const validateRepeatPasswordHandler = (repeatPassword, { password }) => {
-    if (!repeatPassword) return ERROR_MESSAGES.required;
-    if (repeatPassword !== password) return ERROR_MESSAGES.repeatPassword;
-  };
+  const { passwordStrengthLevel, validatePassword, validateRepeatPassword } =
+    usePasswordValidation();
 
   return (
     <AuthPageWrapper title='Set new password'>
       <Formik initialValues={setNewPasswordInitialValues()} onSubmit={() => {}}>
         <Form>
           <div className={styles.fieldsWrapper}>
-            <PasswordField
-              name='password'
-              label='Password'
-              validate={validatePasswordHandler}
-            />
+            <PasswordField name='password' label='Password' validate={validatePassword} />
             <PasswordField
               name='repeatPassword'
               label='Repeat Password'
-              validate={validateRepeatPasswordHandler}
+              validate={validateRepeatPassword}
             />
           </div>
           <PasswordValidator passedLevel={passwordStrengthLevel} />
