+++ src/src_V2/modules/auth/signUp/SignUp.jsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { Formik, Form } from 'formik';
 import { useLocation } from 'react-router-dom';
 
@@ -8,25 +7,15 @@ import AuthPageWrapper from '../components/authWrapper/authPageWrapper';
 import PasswordValidator from '../components/passwordValidator/PasswordValidator';
 
 import styles from './SignUp.module.css';
-import { ERROR_MESSAGES, getPasswordStrengthLevel } from '../../../utils';
+import { usePasswordValidation } from '../../../hooks';
 import { setNewPasswordInitialValues } from '../../../utils/schemas/auth/auth';
 
 export default function SignUp() {
-  const [passwordStrengthLevel, setPasswordStrengthLevel] = useState(0);
+  const { passwordStrengthLevel, validatePassword, validateRepeatPassword } =
+    usePasswordValidation();
+
   const location = useLocation();
   const params = new URLSearchParams(location.search);
-
-  const validatePasswordHandler = (password) => {
-    const strengthLevel = getPasswordStrengthLevel(password);
-    setPasswordStrengthLevel(strengthLevel);
-    if (!password) return ERROR_MESSAGES.required;
-    if (strengthLevel < 31) return ERROR_MESSAGES.password;
-  };
-
-  const validateRepeatPasswordHandler = (repeatPassword, { password }) => {
-    if (!repeatPassword) return ERROR_MESSAGES.required;
-    if (repeatPassword !== password) return ERROR_MESSAGES.repeatPassword;
-  };
 
   return (
     params.get('token') && (
@@ -40,12 +29,12 @@ export default function SignUp() {
               <PasswordField
                 name='password'
                 label='Password'
-                validate={validatePasswordHandler}
+                validate={validatePassword}
               />
               <PasswordField
                 name='repeatPassword'
                 label='Repeat Password'
-                validate={validateRepeatPasswordHandler}
+                validate={validateRepeatPassword}
               />
             </div>
             <PasswordValidator passedLevel={passwordStrengthLevel} />
